+++ src/Annotation/Handler/InjectHandler.php
@@ -9,8 +9,8 @@ use Framework\Annotation\Inject;
 
 return [
     Inject::class => static function (ReflectionProperty $property, mixed $instance, \ReflectionAttribute $self) {
-        $clazz = $property->getType()->getName();
-        if (class_exists($clazz)) {
+        $clazz = $property->getType()?->getName();
+        if ($clazz && class_exists($clazz)) {
             $resource = container($clazz);
             $property->setAccessible(true);
             $property->setValue($instance, $resource);