+++ src/Database/Filter/FilterableTrait.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
  * This file is part of Heros-Worker.
  * @contact  chenzf@pvc123.com
  */
-
 namespace Framework\Database\Filter;
 
 use Illuminate\Database\Eloquent\Builder;
@@ -24,7 +23,7 @@ trait FilterableTrait
     public function scopeFilter(Builder $query, string $filterClazz): Builder
     {
         $filters = new $filterClazz;
-        if (!$filters instanceof BaseFilters) {
+        if (! $filters instanceof BaseFilters) {
             throw new \RuntimeException('filters must extends BaseFilter');
         }
         return $filters->apply($query);