+++ src/Session/RedisClusterSessionHandler.php
@@ -6,25 +6,20 @@ declare(strict_types=1);
  */
 namespace Framework\Session;
 
-use Framework\Exception\HerosException;
-
 class RedisClusterSessionHandler extends RedisSessionHandler
 {
     /**
      * @throws \RedisClusterException
      */
-    public function __construct(array $config)
+    public function __construct($config)
     {
-        if (! extension_loaded('redis')) {
-            throw new HerosException('please install redis ext');
-        }
         parent::__construct($config);
-        $this->_maxLifeTime = (int)ini_get('session.gc_maxlifetime');
+        $this->_maxLifetime = (int)ini_get('session.gc_maxlifetime');
         $timeout = $config['timeout'] ?? 2;
-        $readTimeout = $config['read_timeout'] ?? $timeout;
+        $read_timeout = $config['read_timeout'] ?? $timeout;
         $persistent = $config['persistent'] ?? false;
         $auth = $config['auth'] ?? '';
-        $args = [null, $config['host'], $timeout, $readTimeout, $persistent];
+        $args = [null, $config['host'], $timeout, $read_timeout, $persistent];
         if ($auth) {
             $args[] = $auth;
         }
@@ -33,5 +28,13 @@ class RedisClusterSessionHandler extends RedisSessionHandler
             $config['prefix'] = 'redis_session_';
         }
         $this->_redis->setOption(\Redis::OPT_PREFIX, $config['prefix']);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function read($session_id)
+    {
+        return $this->_redis->get($session_id);
     }
 }