+++ src/Session/RedisSessionHandler.php
@@ -6,15 +6,6 @@ declare(strict_types=1);
  */
 namespace Framework\Session;
 
-use Framework\Exception\HerosException;
-
 class RedisSessionHandler extends \Workerman\Protocols\Http\Session\RedisSessionHandler
 {
-    public function __construct($config)
-    {
-        if (! extension_loaded('redis')) {
-            throw new HerosException('please install redis ext');
-        }
-        parent::__construct($config);
-    }
 }